refactor: extract processing poll from App
